% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAR_cov.R
\name{CAR_cov}
\alias{CAR_cov}
\title{Computes covariance matrix for CAR process}
\usage{
CAR_cov(W, tau, rho)
}
\arguments{
\item{W}{Weight/adjacency matrix (n x n).}

\item{tau}{"Variance" parameter; must be strictly positive.}

\item{rho}{"Correlation" parameter. Must be in \[-1, 1\]}
}
\value{
covariance matrix
}
\description{
Computes covariance matrix for CAR process
}
