% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DreameSpase.R
\name{DreameSpase}
\alias{DreameSpase}
\title{Fits DreameSpase model on given data}
\usage{
DreameSpase(
  data_list,
  X,
  W,
  n_samples,
  n_burnin,
  thinning = 1,
  settings = NULL
)
}
\arguments{
\item{data_list}{List of vectors of outcomes. List entries represent regions,
and vector entries represent outcomes at sub-region levels.}

\item{X}{N x p matrix of covariates, where N is number of regions (i.e. length
of data_list), and p is number of covariates}

\item{W}{List of adjacency matrices. Entry i represents the adjacency matrix
for region i, i.e. length(data_list[i]) = nrow(W[[i]]) = ncol(W[[i]]). These
matrices should be symmetric, and should have 1 in entry i,j if sub-regions
i and j are adjacent, and zeroes otherwise (including the diagonal).}

\item{n_samples}{Number of samples to take after burnin}

\item{n_burnin}{Number of burnin samples}

\item{thinning}{How much if at all to thin the posterior samples. Defaults
to 1, i.e. no thinning}

\item{settings}{Additional settings, including priors. See vignette for
overview of most important settings.}
}
\description{
Fits DreameSpase model on given data
}
