% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RectangularAdjMat.R
\name{RectangularAdjMat}
\alias{RectangularAdjMat}
\title{Generate an adjacency matrix for an m x n lattice}
\usage{
RectangularAdjMat(m, n)
}
\arguments{
\item{m}{number of rows in the lattice}

\item{n}{number of columns in the lattice}
}
\value{
(m\*n) x (m\*n) adjacency matrix for an m x n lattice
}
\description{
Generate an adjacency matrix for an m x n lattice
}
