% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_chol.R
\name{rmvnorm_chol}
\alias{rmvnorm_chol}
\title{Samples from a multivariate normal distribution using cholesky decomposition.}
\usage{
rmvnorm_chol(n, mu, Sigma)
}
\arguments{
\item{n}{Number of samples from specified distribution.}

\item{mu}{Mean vector of distribution (p x 1)}

\item{Sigma}{Covariance matrix of distribution (p x p).}
}
\value{
n x p matrix of samples
}
\description{
Samples from a multivariate normal distribution using cholesky decomposition.
}
