% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCAR.R
\name{SampleCAR}
\alias{SampleCAR}
\title{Samples from standard CAR process.}
\usage{
SampleCAR(W, tau, rho)
}
\arguments{
\item{W}{Weight/adjacency matrix (n x n).}

\item{tau}{"Variance" parameter; must be strictly positive.}

\item{rho}{"Correlation" parameter. Must be in \[-1, 1\]}
}
\value{
numeric vector of length n of samples
}
\description{
Samples from standard CAR process.
}
