% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateDataSet.R
\name{SimulateDataSet}
\alias{SimulateDataSet}
\title{Simulates N sets of data - returns list of simulated data objects.}
\usage{
SimulateDataSet(N, m, n, X_mu, X_sigma, alpha, tau_2, rho, psi_2, nu_2, phi)
}
\value{
List of length N, each entry of which is a list with the following
attributes: `data` is the final simulated lattice; `mu` is the mean component;
`alpha` is the effect of the covariates on the lattice;
`delta` is the spatial error; `epsilon` is the pure error; `X` is the
simulated covariates; `max_val` and `min_val` are the maximum and minimum
values simulated on the lattice, respectively; `W` is the adjacency matrix;
`D_w` is a diagonal matrix whose entries are the row sums of the
corresponding rows of `W` (this is used in sampling from CARs on the lattice).
}
\description{
Simulates N sets of data - returns list of simulated data objects.
}
